+++ packages/clerk-js/src/utils/url.ts
@@ -1,4 +1,4 @@
-import { camelToSnake, isIPV4Address } from '@clerk/shared';
+import { camelToSnake, createDevOrStagingUrlCache, isIPV4Address } from '@clerk/shared';
 import type { SignUpResource } from '@clerk/types';
 
 import { loadScript } from '../utils';
@@ -36,22 +36,9 @@ export const DEV_OR_STAGING_SUFFIXES = [
 
 const BANNED_URI_PROTOCOLS = ['javascript:'] as const;
 
-const devOrStagingUrlCache = new Map<string, boolean>();
+const { isDevOrStagingUrl } = createDevOrStagingUrlCache();
+export { isDevOrStagingUrl };
 const accountsCache = new Map<string, boolean>();
-
-export function isDevOrStagingUrl(url: string | URL): boolean {
-  if (!url) {
-    return false;
-  }
-
-  const hostname = typeof url === 'string' ? url : url.hostname;
-  let res = devOrStagingUrlCache.get(hostname);
-  if (res === undefined) {
-    res = DEV_OR_STAGING_SUFFIXES.some(s => hostname.endsWith(s));
-    devOrStagingUrlCache.set(hostname, res);
-  }
-  return res;
-}
 
 export function isAccountsHostedPages(url: string | URL = window.location.hostname): boolean {
   if (!url) {
@@ -133,7 +120,7 @@ interface BuildURLOptions<T> {
  * References:
  * https://developer.mozilla.org/en-US/docs/Web/API/URL
  *
- * @param {URLParams} params
+ * @param {BuildURLParams} params
  * @param {BuildURLOptions} options
  * @returns {URL | string} Returns the URL href
  */
@@ -143,7 +130,7 @@ export function buildURL<B extends boolean>(
 ): B extends true ? string : URL;
 
 export function buildURL(params: BuildURLParams, options: BuildURLOptions<boolean> = {}): URL | string {
-  const { base, hashPath, hashSearch, ...rest } = params;
+  const { base, hashPath, hashSearch, searchParams, ...rest } = params;
 
   let fallbackBase = '';
   // This check is necessary for React native environments where window is undefined.
@@ -156,6 +143,14 @@ export function buildURL(params: BuildURLParams, options: BuildURLOptions<boolea
   }
 
   const url = new URL(base || '', fallbackBase);
+
+  // Properly copy URLSearchParams
+  if (searchParams instanceof URLSearchParams) {
+    searchParams.forEach((value, key) => {
+      url.searchParams.set(key, value);
+    });
+  }
+
   Object.assign(url, rest);
 
   // Treat that hash part of the main URL as if it's another URL with a pathname and a search.
@@ -214,8 +209,8 @@ export function stripOrigin(url: URL | string): string {
  *
  * Strips the trailing slashes from a string
  *
- * @param {path} string
  * @returns {string} Returns the string without trailing slashes
+ * @param path
  */
 export const trimTrailingSlash = (path: string): string => {
   return (path || '').replace(/\/+$/, '');