+++ packages/types/src/clerk.ts
@@ -69,11 +69,11 @@ export interface Clerk {
   /** Clerk Proxy url string. */
   proxyUrl?: string;
 
-  /**
-   * @experimental
-   * Clerk Satellite Frontend API string.
-   */
-  domain?: string | ((url: URL) => string);
+  /** Clerk Satellite Frontend API string. */
+  domain: string;
+
+  /** Clerk Flag for satellite apps. */
+  isSatellite: boolean;
 
   instanceType?: InstanceType;
 
@@ -472,6 +472,7 @@ export interface ClerkOptions {
    * @experimental
    */
   isSatellite?: boolean | ((url: URL) => boolean);
+  signInUrl?: string;
 }
 
 export interface Resources {